Ensure consistent style and formatting for Solidity code

+++ common/common/src/main/solidity/BytesUtil.sol
@@ -15,42 +15,60 @@
 pragma solidity >=0.7.1;
 
 abstract contract BytesUtil {
-
     // Based on stack overflow here: https://ethereum.stackexchange.com/questions/15350/how-to-convert-an-bytes-to-address-in-solidity
-    function bytesToAddress1(bytes memory _b, uint256 _startOffset) internal pure returns (address addr) {
+    function bytesToAddress1(bytes memory _b, uint256 _startOffset)
+        internal
+        pure
+        returns (address addr)
+    {
         assembly {
             addr := mload(add(_b, add(32, _startOffset)))
         }
     }
 
-    function bytesToAddress2(bytes memory _b, uint256 _startOffset) internal pure returns (address addr) {
+    function bytesToAddress2(bytes memory _b, uint256 _startOffset)
+        internal
+        pure
+        returns (address addr)
+    {
         assembly {
             addr := mload(add(_b, add(20, _startOffset)))
         }
     }
 
-
-    function bytesToAddress(bytes memory _b) internal pure returns (address addr) {
+    function bytesToAddress(bytes memory _b)
+        internal
+        pure
+        returns (address addr)
+    {
         assembly {
             addr := mload(add(_b, 20))
         }
     }
 
-
     // TODO find something faster than this.
     // From stack overflow here: https://ethereum.stackexchange.com/questions/7702/how-to-convert-byte-array-to-bytes32-in-solidity
-    function bytesToBytes32CallData(bytes calldata b, uint offset) internal pure returns (bytes32) {
+    function bytesToBytes32CallData(bytes calldata b, uint256 offset)
+        internal
+        pure
+        returns (bytes32)
+    {
         bytes32 out;
 
-        for (uint i = 0; i < 32; i++) {
+        for (uint256 i = 0; i < 32; i++) {
             out |= bytes32(b[offset + i] & 0xFF) >> (i * 8);
         }
         return out;
     }
-    function bytesToBytes32(bytes memory b, uint offset) internal pure returns (bytes32) {
+
+    function bytesToBytes32(bytes memory b, uint256 offset)
+        internal
+        pure
+        returns (bytes32)
+    {
         bytes32 out;
 
-        for (uint i = 0; i < 32; i++) {
+        for (uint256 i = 0; i < 32; i++) {
             out |= bytes32(b[offset + i] & 0xFF) >> (i * 8);
         }
         return out;
@@ -58,8 +76,15 @@ abstract contract BytesUtil {
 
     // Starting point was this, but with some modifications.
     // https://ethereum.stackexchange.com/questions/49185/solidity-conversion-bytes-memory-to-uint
-    function bytesToUint256(bytes memory _b, uint256 _startOffset) internal pure returns (uint256) {
-        require(_b.length >= _startOffset + 32, "slicing out of range (uint256)");
+    function bytesToUint256(bytes memory _b, uint256 _startOffset)
+        internal
+        pure
+        returns (uint256)
+    {
+        require(
+            _b.length >= _startOffset + 32,
+            "slicing out of range (uint256)"
+        );
         uint256 x;
         assembly {
             x := mload(add(_b, add(32, _startOffset)))
@@ -67,7 +92,11 @@ abstract contract BytesUtil {
         return x;
     }
 
-    function bytesToUint64(bytes memory _b, uint256 _startOffset) internal pure returns (uint64) {
+    function bytesToUint64(bytes memory _b, uint256 _startOffset)
+        internal
+        pure
+        returns (uint64)
+    {
         require(_b.length >= _startOffset + 8, "slicing out of range (uint64)");
         uint256 x;
         assembly {
@@ -76,7 +105,11 @@ abstract contract BytesUtil {
         return uint64(x);
     }
 
-    function bytesToUint32(bytes memory _b, uint256 _startOffset) internal pure returns (uint32) {
+    function bytesToUint32(bytes memory _b, uint256 _startOffset)
+        internal
+        pure
+        returns (uint32)
+    {
         require(_b.length >= _startOffset + 4, "slicing out of range (uint32)");
         uint256 x;
         assembly {
@@ -85,7 +118,11 @@ abstract contract BytesUtil {
         return uint32(x);
     }
 
-    function bytesToUint16(bytes memory _b, uint256 _startOffset) internal pure returns (uint16) {
+    function bytesToUint16(bytes memory _b, uint256 _startOffset)
+        internal
+        pure
+        returns (uint16)
+    {
         require(_b.length >= _startOffset + 2, "slicing out of range (uint16)");
         uint256 x;
         assembly {
@@ -94,7 +131,11 @@ abstract contract BytesUtil {
         return uint16(x);
     }
 
-    function bytesToUint8(bytes memory _b, uint256 _startOffset) internal pure returns (uint8) {
+    function bytesToUint8(bytes memory _b, uint256 _startOffset)
+        internal
+        pure
+        returns (uint8)
+    {
         require(_b.length >= _startOffset + 1, "slicing out of range (uint8)");
         uint256 x;
         assembly {
@@ -104,7 +145,11 @@ abstract contract BytesUtil {
     }
 
     // From https://github.com/GNSPS/solidity-bytes-utils/blob/master/contracts/BytesLib.sol
-    function sliceAsm(bytes memory _bytes, uint256 _start, uint256 _length) internal pure returns (bytes memory) {
+    function sliceAsm(
+        bytes memory _bytes,
+        uint256 _start,
+        uint256 _length
+    ) internal pure returns (bytes memory) {
         require(_bytes.length >= (_start + _length), "Read out of bounds");
 
         bytes memory tempBytes;
@@ -112,31 +157,40 @@ abstract contract BytesUtil {
         assembly {
             switch iszero(_length)
             case 0 {
-            // Get a location of some free memory and store it in tempBytes as
-            // Solidity does for memory variables.
-            tempBytes := mload(0x40)
+                // Get a location of some free memory and store it in tempBytes as
+                // Solidity does for memory variables.
+                tempBytes := mload(0x40)
 
-            // The first word of the slice result is potentially a partial
-            // word read from the original array. To read it, we calculate
-            // the length of that partial word and start copying that many
-            // bytes into the array. The first word we copy will start with
-            // data we don't care about, but the last `lengthmod` bytes will
-            // land at the beginning of the contents of the new array. When
-            // we're done copying, we overwrite the full first word with
-            // the actual length of the slice.
-            let lengthmod := and(_length, 31)
+                // The first word of the slice result is potentially a partial
+                // word read from the original array. To read it, we calculate
+                // the length of that partial word and start copying that many
+                // bytes into the array. The first word we copy will start with
+                // data we don't care about, but the last `lengthmod` bytes will
+                // land at the beginning of the contents of the new array. When
+                // we're done copying, we overwrite the full first word with
+                // the actual length of the slice.
+                let lengthmod := and(_length, 31)
 
-            // The multiplication in the next line is necessary
-            // because when slicing multiples of 32 bytes (lengthmod == 0)
-            // the following copy loop was copying the origin's length
-            // and then ending prematurely not copying everything it should.
-            let mc := add(add(tempBytes, lengthmod), mul(0x20, iszero(lengthmod)))
-            let end := add(mc, _length)
+                // The multiplication in the next line is necessary
+                // because when slicing multiples of 32 bytes (lengthmod == 0)
+                // the following copy loop was copying the origin's length
+                // and then ending prematurely not copying everything it should.
+                let mc := add(
+                    add(tempBytes, lengthmod),
+                    mul(0x20, iszero(lengthmod))
+                )
+                let end := add(mc, _length)
 
                 for {
-                // The multiplication in the next line has the same exact purpose
-                // as the one above.
-                    let cc := add(add(add(_bytes, lengthmod), mul(0x20, iszero(lengthmod))), _start)
+                    // The multiplication in the next line has the same exact purpose
+                    // as the one above.
+                    let cc := add(
+                        add(
+                            add(_bytes, lengthmod),
+                            mul(0x20, iszero(lengthmod))
+                        ),
+                        _start
+                    )
                 } lt(mc, end) {
                     mc := add(mc, 0x20)
                     cc := add(cc, 0x20)
@@ -146,8 +200,8 @@ abstract contract BytesUtil {
 
                 mstore(tempBytes, _length)
 
-            //update free-memory pointer
-            //allocating the array padded to 32 bytes like the compiler does now
+                //update free-memory pointer
+                //allocating the array padded to 32 bytes like the compiler does now
                 mstore(0x40, and(add(mc, 31), not(31)))
             }
             //if we want a zero-length slice let's just return a zero-length array
@@ -162,20 +216,32 @@ abstract contract BytesUtil {
     }
 
     // https://ethereum.stackexchange.com/questions/78559/how-can-i-slice-bytes-strings-and-arrays-in-solidity
-    function slice(bytes memory _bytes, uint256 _start, uint256 _length) internal pure returns (bytes memory) {
+    function slice(
+        bytes memory _bytes,
+        uint256 _start,
+        uint256 _length
+    ) internal pure returns (bytes memory) {
         bytes memory a = new bytes(_length);
-        for(uint i=0; i < _length; i++){
-            a[i] = _bytes[_start+i];
+        for (uint256 i = 0; i < _length; i++) {
+            a[i] = _bytes[_start + i];
         }
         return a;
     }
 
-    function slice(bytes memory _bytes, uint256 _start) internal pure returns (bytes memory) {
+    function slice(bytes memory _bytes, uint256 _start)
+        internal
+        pure
+        returns (bytes memory)
+    {
         return slice(_bytes, _start, (_bytes.length - _start));
     }
 
-    function compare(bytes memory _a, bytes memory _b) internal pure returns (bool) {
-        if(_a.length != _b.length) {
+    function compare(bytes memory _a, bytes memory _b)
+        internal
+        pure
+        returns (bool)
+    {
+        if (_a.length != _b.length) {
             return false;
         } else {
             return keccak256(_a) == keccak256(_b);
